Move Custom User Menu to its own file
