+++ Marlin/src/lcd/menu/menu.cpp
@@ -388,63 +388,6 @@ void line_to_z(const float &z) {
   line_to_current_z();
 }
 
-#if ENABLED(CUSTOM_USER_MENUS)
-
-  #ifdef USER_SCRIPT_DONE
-    #define _DONE_SCRIPT "\n" USER_SCRIPT_DONE
-  #else
-    #define _DONE_SCRIPT ""
-  #endif
-
-  void _lcd_user_gcode(PGM_P const cmd) {
-    enqueue_and_echo_commands_P(cmd);
-    #if ENABLED(USER_SCRIPT_AUDIBLE_FEEDBACK)
-      lcd_completion_feedback();
-    #endif
-    #if ENABLED(USER_SCRIPT_RETURN)
-      lcd_return_to_status();
-    #endif
-  }
-
-  #if defined(USER_DESC_1) && defined(USER_GCODE_1)
-    void lcd_user_gcode_1() { _lcd_user_gcode(PSTR(USER_GCODE_1 _DONE_SCRIPT)); }
-  #endif
-  #if defined(USER_DESC_2) && defined(USER_GCODE_2)
-    void lcd_user_gcode_2() { _lcd_user_gcode(PSTR(USER_GCODE_2 _DONE_SCRIPT)); }
-  #endif
-  #if defined(USER_DESC_3) && defined(USER_GCODE_3)
-    void lcd_user_gcode_3() { _lcd_user_gcode(PSTR(USER_GCODE_3 _DONE_SCRIPT)); }
-  #endif
-  #if defined(USER_DESC_4) && defined(USER_GCODE_4)
-    void lcd_user_gcode_4() { _lcd_user_gcode(PSTR(USER_GCODE_4 _DONE_SCRIPT)); }
-  #endif
-  #if defined(USER_DESC_5) && defined(USER_GCODE_5)
-    void lcd_user_gcode_5() { _lcd_user_gcode(PSTR(USER_GCODE_5 _DONE_SCRIPT)); }
-  #endif
-
-  void _menu_user() {
-    START_MENU();
-    MENU_BACK(MSG_MAIN);
-    #if defined(USER_DESC_1) && defined(USER_GCODE_1)
-      MENU_ITEM(function, USER_DESC_1, lcd_user_gcode_1);
-    #endif
-    #if defined(USER_DESC_2) && defined(USER_GCODE_2)
-      MENU_ITEM(function, USER_DESC_2, lcd_user_gcode_2);
-    #endif
-    #if defined(USER_DESC_3) && defined(USER_GCODE_3)
-      MENU_ITEM(function, USER_DESC_3, lcd_user_gcode_3);
-    #endif
-    #if defined(USER_DESC_4) && defined(USER_GCODE_4)
-      MENU_ITEM(function, USER_DESC_4, lcd_user_gcode_4);
-    #endif
-    #if defined(USER_DESC_5) && defined(USER_GCODE_5)
-      MENU_ITEM(function, USER_DESC_5, lcd_user_gcode_5);
-    #endif
-    END_MENU();
-  }
-
-#endif
-
 #if ENABLED(BABYSTEP_ZPROBE_OFFSET)
 
   void lcd_babystep_zoffset() {
